HaProxy: Change `drain` state name to `drained`
This (a) avoids a namespace collision between the function called when the state
is set to `drained` and the `drain` module parameter and (b) changes the tense
of the former `drain` state to match that of `enabled` and `disabled`. The
past-tense also makes more sense from the perspective that we're declaring the
desired state of a host.